clientv3: remove strict yaml dependency
Moved to clientv3/yaml
